test(api): adapt fetch tests to check for xorurl field
