+++ resources/test-scripts/all-tests
@@ -64,6 +64,8 @@ function run_cli_tests() {
     rm -f fake_vault_data.json
     cargo test --release --features=scl-mock --test cli_cat -- --test-threads=1
     rm -f fake_vault_data.json
+    cargo test --release --features=scl-mock --test cli_dog -- --test-threads=1
+    rm -f fake_vault_data.json
     cargo test --release --features=scl-mock --test cli_files -- --test-threads=1
     rm -f fake_vault_data.json
     cargo test --release --features=scl-mock --test cli_keys -- --test-threads=1
@@ -82,6 +84,7 @@ function run_cli_tests_with_auth_daemon() {
     cargo run --release --features=mock-network -- auth --port $RANDOM_PORT_NUMBER
 
     cargo test --release --features=mock-network --test cli_cat -- --test-threads=1
+    cargo test --release --features=mock-network --test cli_dog -- --test-threads=1
     cargo test --release --features=mock-network --test cli_files -- --test-threads=1
     cargo test --release --features=mock-network --test cli_keys -- --test-threads=1
     cargo test --release --features=mock-network --test cli_wallet -- --test-threads=1
